Event Section: Search Bar and Alert when Appointment of other user is adjusted

+++ src/components/MyCalendarComponent.js
@@ -7,6 +7,13 @@ import TextField from '@material-ui/core/TextField';
 import MenuItem from '@material-ui/core/MenuItem';
 import Autocomplete from "@material-ui/lab/Autocomplete";
 import Chip from "@material-ui/core/Chip";
+import AppointmentAlertComponent from "./AppointmentAlertComponent";
+import Alert from '@material-ui/lab/Alert';
+import IconButton from '@material-ui/core/IconButton';
+import Button from '@material-ui/core/Button';
+import CloseIcon from '@material-ui/icons/Close';
+import Collapse from "@material-ui/core/Collapse";
+import Box from "@material-ui/core/Box";
 
 import {
     Scheduler,
@@ -22,12 +29,13 @@ import {
     TodayButton,
 
 } from '@devexpress/dx-react-scheduler-material-ui';
-import {withStyles} from "@material-ui/core/styles";
+import {makeStyles, withStyles} from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import UserService from "../services/UserService";
 import MatchService from "../services/MatchService";
 import {useSelector} from "react-redux";
 import {useEffect} from "react";
+import App from "../App";
 
 
 const style = ({palette}) => ({
@@ -61,18 +69,19 @@ const messages = {
 };
 
 const TextEditor = (props) => {
-    if(props.type === 'multilineTextEditor'){
+    if (props.type === 'multilineTextEditor') {
         return null;
-    }return <AppointmentForm.TextEditor {...props} />
+    }
+    return <AppointmentForm.TextEditor {...props} />
 }
 
 const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
 
     // const [allInterests, setAllInterests] = React.useState([]);
 
-    const allInterests = ["Business and Industry","Advertising","Agriculture","Architecture","Aviation","Banking","Investment banking",
-        "Online banking","Retail banking","Business","Construction","Design","Fashion design","Graphic design","Interior design","Economics","Engineering", "Entrepreneurship", "Health care",
-        "Higher education", "Management", "Marketing", "Nursing", "Online","Digital marketing",
+    const allInterests = ["Business and Industry", "Advertising", "Agriculture", "Architecture", "Aviation", "Banking", "Investment banking",
+        "Online banking", "Retail banking", "Business", "Construction", "Design", "Fashion design", "Graphic design", "Interior design", "Economics", "Engineering", "Entrepreneurship", "Health care",
+        "Higher education", "Management", "Marketing", "Nursing", "Online", "Digital marketing",
         "Display advertising", "Email marketing", "Online advertising", "Search engine optimization", "Social media", "Social media marketing",
         "Web design", "Web development", "Web hosting", "Personal finance", "Creditcards", "Insurance", "Investment", "Mortgage loans", "Real estate",
         "Retail", "Sales", "Science", "Small business", "Entertainment", "Games", "Action games", "Board games",
@@ -80,7 +89,7 @@ const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
         "Online games", "Online poker", "Puzzle video games", "Racing games", "Role-playing games", "Shooter games", "Simulation games", "Sports games", "Strategy games", "Video games",
         "Word games", "Live events", "Ballet", "Bars", "Concerts", "Dancehalls", "Music festivals", "Nightclubs", "Parties", "Plays", "Theatre", "Movies", "Action movies",
         "Animated movies", "Anime movies", "Bollywood movies", "Comedy movies", "Documentary movies", "Drama movies", "Fantasy movies", "Horror movies", "Musical theatre", "Science fiction movies",
-        "Thriller movies", "Music", "Blues music","Classical music", "Country music", "Dance music", "Electronic music", "Gospel music",
+        "Thriller movies", "Music", "Blues music", "Classical music", "Country music", "Dance music", "Electronic music", "Gospel music",
         "Heavy metal music", "Hip hop music", "Jazz music", "Music videos", "Pop music", "Rhythm and blues music", "Rock musicSoul music", "Reading", "Books",
         "Comics", "E-books", "Fiction books", "Literature", "Magazines", "Manga", "Mystery fiction", "Newspapers",
         "Non-fiction books", "Romance novels", "TV", "TV comedies", "TV game shows", "TV reality shows", "TV talkshows", "Family and relationships",
@@ -91,17 +100,17 @@ const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
         "Mexican cuisine", "Middle Eastern cuisine", "Spanish cuisine", "Thai cuisine", "Vietnamese cuisine", "Food", "Barbecue", "Chocolate", "Desserts", "Fast food",
         "Organic food", "Pizza", "Seafood", "Veganism", "Vegetarianism", "Restaurants", "Coffeehouses",
         "Diners", "Fast casual restaurants", "Fast food restaurants", "Hobbies and activities", "Arts and music", "Acting",
-        "Crafts", "Dance","Drawing", "Drums", "Fine art", "Guitar", "Painting", "Performing arts", "Photography","Sculpture", "Singing", "Writing", "Current events",
+        "Crafts", "Dance", "Drawing", "Drums", "Fine art", "Guitar", "Painting", "Performing arts", "Photography", "Sculpture", "Singing", "Writing", "Current events",
         "Home and garden", "Do it yourself (DIY)", "Furniture", "Gardening", "Home Appliances", "Home improvement", "Pets", "Birds", "Cats", "Dogs", "Fish", "Horses", "Pet food",
-        "Rabbits", "Reptiles", "Politics and social issues","Charity and causes", "Community issues",
-        "Environmentalism","Law", "Military","Politics", "Religion","Sustainability", "Veterans","Volunteering", "Travel", "Adventure travel","Air travel","Beaches",
+        "Rabbits", "Reptiles", "Politics and social issues", "Charity and causes", "Community issues",
+        "Environmentalism", "Law", "Military", "Politics", "Religion", "Sustainability", "Veterans", "Volunteering", "Travel", "Adventure travel", "Air travel", "Beaches",
         "Car rentals", "Cruises", "Ecotourism", "Hotels", "Lakes", "Mountains", "Nature", "Theme parks", "Tourism", "Vacations",
         "Vehicles", "Automobiles", "Boats", "Electric vehicle", "Hybrids", "Minivans", "Motorcycles", "RVs", "SUVs", "Scooters", "Trucks", "Shopping and fashion", "Beauty", "Beauty salons", "Cosmetics", "Fragrances",
         "Hair products", "Spas", "Tattoos", "Clothing", "Children’s clothing", "Men’s clothing", "Shoes", "Women’s clothing", "Fashionaccessories", "Dresses", "Handbags", "Jewelry",
-        "Sunglasses","Shopping", "Boutiques", "Coupons", "Discount stores", "Luxury goods", "Online shopping", "Shopping malls", "Toys", "Sports and outdoors", "Outdoor recreation", "Boating", "Camping", "Fishing",
+        "Sunglasses", "Shopping", "Boutiques", "Coupons", "Discount stores", "Luxury goods", "Online shopping", "Shopping malls", "Toys", "Sports and outdoors", "Outdoor recreation", "Boating", "Camping", "Fishing",
         "Horseback riding", "Hunting", "Mountain biking", "Surfing", "Sports", "American football", "Association football (Soccer)", "Auto racing", "Baseball", "Basketball", "College football", "Golf", "Marathons",
-        "Skiing", "Snowboarding","Swimming", "Tennis", "Thriathlons", "Volleyball", "Technology", "Computers", "Computer memory", "Computer monitors", "Computer processors", "Computer servers", "Desktop computers", "Free software", "Hard drives", "Network storage", "Software", "Tablet computers",
-        "Consumer electronics","Audio equipment", "Camcorders", "Cameras", "E-book readers", "GPS devices", "Game consoles", "Mobile phones", "Portable media players", "Projectors", "Smartphones", "Televisions",
+        "Skiing", "Snowboarding", "Swimming", "Tennis", "Thriathlons", "Volleyball", "Technology", "Computers", "Computer memory", "Computer monitors", "Computer processors", "Computer servers", "Desktop computers", "Free software", "Hard drives", "Network storage", "Software", "Tablet computers",
+        "Consumer electronics", "Audio equipment", "Camcorders", "Cameras", "E-book readers", "GPS devices", "Game consoles", "Mobile phones", "Portable media players", "Projectors", "Smartphones", "Televisions",
     ];
     // const extractInts = async() => {
     //     let ints = await MatchService.getInterests();
@@ -126,23 +135,23 @@ const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
 
     return (
         <AppointmentForm.BasicLayout
-           appointmentData={appointmentData}
-           onFieldChange={onFieldChange}
+            appointmentData={appointmentData}
+            onFieldChange={onFieldChange}
             {...restProps}
-            >
+        >
             <AppointmentForm.Label
-                text = "Video URL"
-                type = "title"
-                />
+                text="Video URL"
+                type="title"
+            />
             <AppointmentForm.TextEditor
                 value={appointmentData.link}
                 placeholder="We will automatically set this value after saving :)"
                 onValueChange={onURLFieldChange}
                 readOnly
-                />
+            />
             <AppointmentForm.Label
-                text = "Description"
-                type = "title"
+                text="Description"
+                type="title"
             />
             <AppointmentForm.TextEditor
                 type='multilineTextEditor'
@@ -152,9 +161,9 @@ const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
             />
 
             <AppointmentForm.Label
-                style={{'marginTop':'20px'}}
-                text = "Interests"
-                type = "title"
+                style={{'marginTop': '20px'}}
+                text="Interests"
+                type="title"
             />
 
             <Autocomplete
@@ -179,8 +188,17 @@ const BasicLayout = ({onFieldChange, appointmentData, ...restProps}) => {
     );
 };
 
+const styles = theme => ({
+    root: {
+        width: '100%',
+        '& > * + *': {
+            marginTop: theme.spacing(2),
+        },
+    },
+});
 
-export default class MyCalendarComponent extends React.PureComponent {
+
+class MyCalendarComponent extends React.PureComponent {
 
 
     constructor(props) {
@@ -192,7 +210,9 @@ export default class MyCalendarComponent extends React.PureComponent {
             appointmentChanges: {},
             editingAppointment: undefined,
             user: props.user,
+            unallowedEdit: false,
             mappings: {},
+            open: true,
         };
         this.currentDataChange = (currentDate) => {
             this.setState({currentDate});
@@ -211,7 +231,7 @@ export default class MyCalendarComponent extends React.PureComponent {
         description: appointment.description,
         link: appointment.link,
         user: appointment.user,
-        interests:appointment.interests,
+        interests: appointment.interests,
     });
 
     async componentDidMount() {
@@ -226,14 +246,18 @@ export default class MyCalendarComponent extends React.PureComponent {
         let month = await newDate.getMonth() + 1;
         let year = await newDate.getFullYear();
 
-        let nDate = ""+year+"-"+month+"-"+date;
+        let nDate = "" + year + "-" + month + "-" + date;
 
         console.log(" date: " + date);
 
         let allInterests = await MatchService.getInterests()
-        this.setState({data: res_mapped, mapping: mapping, allInterests: allInterests[0].facebookInterests, currentDate: nDate});
+        this.setState({
+            data: res_mapped,
+            mapping: mapping,
+            allInterests: allInterests[0].facebookInterests,
+            currentDate: nDate
+        });
     }
-
 
 
     changeAddedAppointment(addedAppointment) {
@@ -248,57 +272,92 @@ export default class MyCalendarComponent extends React.PureComponent {
         this.setState({editingAppointment});
     }
 
+    setUnallowedEdit(value) {
+        this.setState({unallowedEdit: value})
+    }
+
     async commitChanges({added, changed, deleted}) {
         var app = null;
 
-        if (added){
-            try{
+        if (added) {
+            try {
                 added["user"] = this.state.user._id;
-                added["link"] = "localhost:3000/call/"+Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
+                added["link"] = "localhost:3000/call/" + Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
                 const response = await AppointmentService.createAppointment(added);
                 app = response.map(this.mapAppointmentData)[0];
-            }catch (e) {
+            } catch (e) {
+                return e;
+            }
+        }
+
+        if (changed || deleted) {
+            try {
+                var key=0;
+                if (changed){
+                    key = Object.keys(changed)[0];
+                }else{
+                    key = deleted;
+                }
+
+                const response = await AppointmentService.getAppointment(key);
+                app = response;
+            } catch (e) {
                 return e;
             }
         }
         this.setState((state) => {
-            let {data} = state;
+            let {data, unallowedEdit} = state;
             if (added) {
                 //const startingAddedId = data.length > 0 ? data[data.length - 1].id + 1 : 0;
                 data = [...data, {id: app.id, ...added}];
             }
             if (changed) {
 
-                data = data.map(appointment => (
-                    changed[appointment.id] ? {...appointment, ...changed[appointment.id]} : appointment));
-                const key = Object.keys(changed)[0];
-                var obj = null;
-                for(var i in data){
-                    if(data[i].id == key){
-                        // Add an element to the dictionary
-                        obj = data[i];
-                        const resp = AppointmentService.updateAppointment(obj);
-                        console.log("resp: " + resp);
-                        break; // If you want to break out of the loop once you've found a match
-                    }
-                }
+                if (app.user == this.state.user._id) {
 
+                    console.log("Getting in here");
+                    data = data.map(appointment => (
+                        changed[appointment.id] ? {...appointment, ...changed[appointment.id]} : appointment));
+                    const key = Object.keys(changed)[0];
+
+                    var obj = null;
+                    for (var i in data) {
+                        if (data[i].id == key) {
+                            // Add an element to the dictionary
+                            obj = data[i];
+                            const resp = AppointmentService.updateAppointment(obj);
+                            console.log("resp: " + resp);
+                            break; // If you want to break out of the loop once you've found a match
+                        }
+                    }
+                } else {
+                    unallowedEdit = true;
+                    console.log("unallowedEdit: " + this.state.unallowedEdit);
+                }
 
 
             }
             if (deleted !== undefined) {
-                data = data.filter(appointment => appointment.id !== deleted);
-                AppointmentService.deleteAppointment(deleted);
+                if (app.user == this.state.user._id) {
+                    data = data.filter(appointment => appointment.id !== deleted);
+                    AppointmentService.deleteAppointment(deleted);
+                }else{
+                    unallowedEdit = true;
+                }
             }
             console.log(data);
-            return {data};
+            return {data, unallowedEdit};
         });
     }
+
 
     render() {
         const {
             currentDate, data, addedAppointment, appointmentChanges, editingAppointment,
         } = this.state;
+
+
+        const {classes} = this.props;
 
         const Content = withStyles(style, {name: 'Content'})(({
                                                                   children, appointmentData, classes, ...restProps
@@ -309,13 +368,13 @@ export default class MyCalendarComponent extends React.PureComponent {
                         <span>Link: </span>
                     </Grid>
                     <Grid item xs={10}>
-                    {
-                        appointmentData.link ?
-                            (
-                                <Link to={appointmentData.link.split("3000")[1]}>{appointmentData.link}</Link>
-                            ):
-                            (<Link to="#">{appointmentData.link}</Link>)
-                    }
+                        {
+                            appointmentData.link ?
+                                (
+                                    <Link to={appointmentData.link.split("3000")[1]}>{appointmentData.link}</Link>
+                                ) :
+                                (<Link to="#">{appointmentData.link}</Link>)
+                        }
                     </Grid>
                     <Grid item xs={2} className={classes.textCenter}>
                         <span>What: </span>
@@ -340,8 +399,8 @@ export default class MyCalendarComponent extends React.PureComponent {
         ));
 
         return (
-            <div >
-                <Paper >
+            <div>
+                <Paper>
                     <Scheduler
                         data={data}
                         height={660}
@@ -382,7 +441,39 @@ export default class MyCalendarComponent extends React.PureComponent {
                     </Scheduler>
                 </Paper>
 
+                <div>
+                    {this.state.unallowedEdit ? (
+                        <Box pt={5}>
+                            <div className={classes.root}>
+                                <Collapse in={this.state.open}>
+                                    <Alert severity="error"
+                                           action={
+                                               <IconButton
+                                                   aria-label="close"
+                                                   color="inherit"
+                                                   size="small"
+                                                   onClick={() => {
+                                                       this.setState({unallowedEdit: false});
+                                                   }}
+                                               >
+                                                   <CloseIcon
+                                                       fontSize="inherit"/>
+                                               </IconButton>
+                                           }
+                                    >
+                                        You accidently tried to edit or delete an appointment that is not yours...
+                                    </Alert>
+                                </Collapse>
+                            </div>
+                        </Box>
+
+
+                    ) : null}
+                </div>
+
             </div>
         );
     }
 }
+
+export default withStyles(styles)(MyCalendarComponent);
